feat: Support for FontIcon.MirroredWhenRightToLeft
Refactoring of FontIcon code and support for MirroredWhenRightToLeft property.
